requirements.txt
langchain
langchain-astradb
langchain-huggingface
langchain-groq
langchain-community
pypdf
python-dotenv
datasets
pandas
flask
prometheus_client